The program now check if the input sudoku is valid
The program now check if the that the user put is valid. This is a better way that keeping a copy of the solved sudoku, because know when we remove some of the cells the sudoku can have more than one solution and the program will still think this solution is valid. The changes where made to python and c++